Factor clear button semantics into a ClearableWidget comp.
